// frontend/src/api-ms-cv-v2/js-client/api.ts
export interface SectionResponseDTO {
    'id'?: number;
    'type'?: string;
    'htmlContent'?: string;
    'ordre'?: number;
    'createdAt'?: string;
    'updatedAt'?: string;
}
export interface TemplateRequestDTO {
    'nom'?: string;
    'isGlobal'?: boolean;
    'sections'?: Array<SectionRequestDTO>;
}
export interface TemplateResponseDTO {
    'id'?: number;
    'nom'?: string;
    'isGlobal'?: boolean;
    'userId'?: number;
    'sections'?: Array<SectionResponseDTO>;
    'createdAt'?: string;
    'updatedAt'?: string;
}